Move pad types to their own file.
